DFR-908: Contested-Document handling Manage bundle event
