+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/ContestedOrderController.java
@@ -1,5 +1,7 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.controllers;
 
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
@@ -15,12 +17,23 @@ import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
+import uk.gov.hmcts.reform.finrem.caseorchestration.error.InvalidCaseDataException;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.HearingUploadBundleData;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.ContestedCaseOrderService;
 
 import javax.validation.constraints.NotNull;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstants.AUTHORIZATION_HEADER;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.HEARING_DATE;
+import static uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.CCDConfigConstant.HEARING_UPLOAD_BUNDLE_COLLECTION;
 
 @Slf4j
 @RestController
@@ -28,6 +41,7 @@ import static uk.gov.hmcts.reform.finrem.caseorchestration.OrchestrationConstant
 @RequestMapping(value = "/case-orchestration")
 public class ContestedOrderController implements BaseController {
 
+    private final ObjectMapper objectMapper;
     private final ContestedCaseOrderService contestedCaseOrderService;
 
     @PostMapping(path = "/contested/send-order", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
@@ -54,5 +68,72 @@ public class ContestedOrderController implements BaseController {
         log.info("Finished sending contested order for case {}", caseDetails.getId());
 
         return ResponseEntity.ok(AboutToStartOrSubmitCallbackResponse.builder().data(caseDetails.getData()).build());
+    }
+
+    @PostMapping(path = "/contested/validateHearingDate", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
+    @ApiOperation(value = "check hearing date")
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "Callback was processed successfully or in case of an error message is attached to the case",
+            response = AboutToStartOrSubmitCallbackResponse.class),
+        @ApiResponse(code = 400, message = "Bad Request"),
+        @ApiResponse(code = 500, message = "Internal Server Error")})
+    public ResponseEntity<AboutToStartOrSubmitCallbackResponse> checkHearingDate(
+        @NotNull @RequestBody @ApiParam("CaseData") CallbackRequest callback) {
+
+        CaseDetails caseDetails = callback.getCaseDetails();
+        log.info("Received request to check for manage bundle hearing date for Case ID: {}", caseDetails.getId());
+        validateCaseData(callback);
+
+        final List<String> errors = new ArrayList<>();
+
+        try {
+            validateHearingDate(callback);
+        } catch (InvalidCaseDataException invalidCaseDataException) {
+            errors.add(invalidCaseDataException.getMessage());
+        }
+        return ResponseEntity.ok(AboutToStartOrSubmitCallbackResponse
+            .builder()
+            .data(caseDetails.getData())
+            .errors(errors)
+            .build());
+    }
+
+    @PostMapping(path = "/contested/sortUploadedHearingBundles", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
+    @ApiOperation(value = "Documents to be viewed in order of newest first at top of the list")
+    @ApiResponses(value = {
+        @ApiResponse(code = 200, message = "Callback was processed successfully or in case of an error message is attached to the case",
+            response = AboutToStartOrSubmitCallbackResponse.class),
+        @ApiResponse(code = 400, message = "Bad Request"),
+        @ApiResponse(code = 500, message = "Internal Server Error")})
+    public ResponseEntity<AboutToStartOrSubmitCallbackResponse> sortHearingBundles(
+        @NotNull @RequestBody @ApiParam("CaseData") CallbackRequest callback) {
+
+        CaseDetails caseDetails = callback.getCaseDetails();
+        log.info("Received request for doc newest first at top of the list for Case ID: {}", caseDetails.getId());
+        validateCaseData(callback);
+        Map<String, Object> caseData = caseDetails.getData();
+
+        List<HearingUploadBundleData> hearingBundleDataList = Optional.ofNullable(caseData.get(HEARING_UPLOAD_BUNDLE_COLLECTION))
+            .map(this::convertToHearingBundleDataList).orElse(Collections.emptyList());
+
+        hearingBundleDataList.sort((e1, e2) -> e2.getValue().getBundleUploadDate().compareTo(e1.getValue().getBundleUploadDate()));
+
+        caseData.put(HEARING_UPLOAD_BUNDLE_COLLECTION, hearingBundleDataList);
+        return ResponseEntity.ok(AboutToStartOrSubmitCallbackResponse
+            .builder()
+            .data(caseData)
+            .build());
+    }
+
+    private List<HearingUploadBundleData> convertToHearingBundleDataList(Object object) {
+        return objectMapper.convertValue(object, new TypeReference<>() {
+        });
+    }
+
+    private void validateHearingDate(CallbackRequest callbackRequest) {
+        if (callbackRequest.getCaseDetails().getData().get(HEARING_DATE) == null) {
+            log.info("Hearing date for Case ID: {} not found", callbackRequest.getCaseDetails().getId());
+            throw new InvalidCaseDataException(BAD_REQUEST.value(), "Missing hearing date.");
+        }
     }
 }
