+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/HearingBundle.java
@@ -1,16 +1,23 @@
 package uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.time.LocalDate;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
 @Builder
-@Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class HearingBundle {
-    @JsonProperty("hearingUploadBundle")
-    private CaseDocument hearingUploadBundle;
+
+    @JsonProperty("bundleDocuments")
+    private CaseDocument bundleDocuments;
+    @JsonProperty("bundleUploadDate")
+    private LocalDate bundleUploadDate;
 }
