+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/controllers/CaseDataControllerTest.java
@@ -11,8 +11,6 @@ import uk.gov.hmcts.reform.finrem.caseorchestration.service.FeatureToggleService
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.IdamService;
 import uk.gov.hmcts.reform.finrem.caseorchestration.service.UpdateSolicitorDetailsService;
 
-import static com.jayway.jsonassert.impl.matcher.IsCollectionWithSize.hasSize;
-import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.is;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -37,7 +35,6 @@ public class CaseDataControllerTest extends BaseControllerTest {
 
     private static final String CONTESTED_HWF_JSON = "/fixtures/contested/hwf.json";
     private static final String CONTESTED_VALIDATE_HEARING_SUCCESSFULLY_JSON = "/fixtures/contested/validate-hearing-successfully.json";
-    private static final String CONTESTED_VALIDATE_HEARING_DATE_JSON = "/fixtures/contested/manage-bundle-validate-hearing-date.json";
 
     @Autowired private CaseDataController caseDataController;
 
@@ -294,55 +291,5 @@ public class CaseDataControllerTest extends BaseControllerTest {
                 .contentType(APPLICATION_JSON_VALUE))
             .andExpect(status().isOk())
             .andExpect(jsonPath("$.data.civilPartnership", is(NO_VALUE)));
-    }
-
-    @Test
-    public void shouldThrowExceptionWhenHearingDateNotFound() throws Exception {
-        when(idamService.isUserRoleAdmin(isA(String.class))).thenReturn(false);
-        when(caseDataService.isContestedApplication(any())).thenReturn(true);
-
-        loadRequestContentWith(CONTESTED_VALIDATE_HEARING_DATE_JSON);
-        mvc.perform(post("/case-orchestration//contested/validateHearingDate")
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.errors").isArray())
-            .andExpect(jsonPath("$.errors", hasSize(1)))
-            .andExpect(jsonPath("$.errors", hasItem("Missing hearing date.")));
-    }
-
-    @Test
-    public void shouldSuccessfullyProcessWhenHearingDateFound() throws Exception {
-        when(idamService.isUserRoleAdmin(isA(String.class))).thenReturn(false);
-        when(caseDataService.isContestedApplication(any())).thenReturn(true);
-
-        loadRequestContentWith(CONTESTED_VALIDATE_HEARING_SUCCESSFULLY_JSON);
-        mvc.perform(post("/case-orchestration//contested/validateHearingDate")
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.data.hearingDate", is("2019-05-04")));
-    }
-
-    @Test
-    public void reArrangeUploadedBundle() throws Exception {
-        when(idamService.isUserRoleAdmin(isA(String.class))).thenReturn(false);
-        when(caseDataService.isContestedApplication(any())).thenReturn(true);
-
-        loadRequestContentWith(CONTESTED_VALIDATE_HEARING_DATE_JSON);
-        mvc.perform(post("/case-orchestration//contested/sortUploadedHearingBundles")
-                .content(requestContent.toString())
-                .header(AUTHORIZATION_HEADER, AUTH_TOKEN)
-                .contentType(APPLICATION_JSON_VALUE))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.data.hearingUploadBundle").isArray())
-            .andExpect(jsonPath("$.data.hearingUploadBundle[0].value.bundleDocuments.document_filename",
-                is("InterimHearingNotice-1649341720076259.pdf")))
-            .andExpect(jsonPath("$.data.hearingUploadBundle[1].value.bundleDocuments.document_filename",
-                is("dummy1-1649341720076259.pdf")))
-            .andExpect(jsonPath("$.data.hearingUploadBundle[2].value.bundleDocuments.document_filename",
-                is("BulkPrintCoverSheet-1649341720076259.pdf")));
     }
 }
